+++ packages/flusmic/lib/src/flusmic_repository.dart
@@ -89,15 +89,15 @@ class Flusmic {
   ///Get the API root document of prismic repository
   ///Contains all the documents.
   Future<FlusmicResponse> getRootDocument(
-          {String language, String authToken}) async =>
-      await query([], authToken: authToken, language: language);
+          {String language, String authToken, int page = 1}) async =>
+      await query([], authToken: authToken, language: language, page: page);
 
   /// Fetch documents by type
   /// Get all the documents by [type] using the slug.
   Future<FlusmicResponse> getDocumentsByType(String slug,
-          {String language, String authToken}) async =>
+          {String language, String authToken, int page = 1}) async =>
       await query([Predicate.at(DefaultPredicatePath.type(), slug)],
-          authToken: authToken, language: language);
+          authToken: authToken, language: language, page: page);
 
   /// Fetch document by id
   /// Get a documents by [id].
@@ -204,7 +204,12 @@ class Flusmic {
           '&q=[[date.hour-before(${p.path.toString()}, ${p.hour})]]');
 
   String _generateOrdering(Ordering ordering) {
-    final descending = ordering.descending ? ' des' : '';
-    return 'my.${ordering.customType}.${ordering.field}$descending';
+    final descending = ordering.descending ? ' desc' : '';
+    return ordering.map(
+      document: (o) => 'document.${o.type}$descending',
+      firstPublicationDate: (o) => 'document.first_publication_date$descending',
+      lastPublicationDate: (o) => 'document.last_publication_date$descending',
+      type: (o) => 'my.${o.customType}.${o.field}$descending',
+    );
   }
 }